v0.0.13
- New shard.
- Interface is updated.
- Bug fix.

+++ Data/Scripts/Global/Inventory/Shard.cs
@@ -6,15 +6,15 @@ public partial class Shard : Item
 {
     [Export] public string ShardType { get; set; }
     [Export] public int Health { get; set; }
-    [Export] public int Damage { get; set; }
+    [Export] public float Damage { get; set; }
     [Export] public int Speed { get; set; }
     [Export] public float TimeReload { get; set; }
     [Export] public float CritChance { get; set; }
-    [Export] public float MaxRange { get; set; }
+    [Export] public int MaxRange { get; set; }
 
     public Shard() : base() { }
 
-    public Shard(int id, int maxCount, string itemName, string description, string shard2D, int health, int damage, int speed, float timeReload, float critChance, float maxRange) : base(id, maxCount, itemName, description)
+    public Shard(int id, int maxCount, string itemName, string description, string shard2D, int health, float damage, int speed, float timeReload, float critChance, int maxRange) : base(id, maxCount, itemName, description)
     {
         ShardType = shard2D;
         Health = health;
